Added the makefile for the module_00

+++ 42/cpp_module_00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: ochoumou <ochoumou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/27 16:15:11 by ochoumou          #+#    #+#             */
-/*   Updated: 2022/10/09 10:59:03 by ochoumou         ###   ########.fr       */
+/*   Updated: 2022/10/10 14:50:14 by ochoumou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int main(int argc, char **argv)
 		for (int i = 1; i < argc; i++)
 		{
 			arg = argv[i];
-			for (int j = 0; j < arg.length(); j++)
+			for (int j = 0; j < (int)arg.length(); j++)
 				std::cout << (char)std::toupper(argv[i][j]);
 		}
 		std::cout << std::endl;
